requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
